refactor: re-organize state
