+++ src/commands/prompt/chat-commands.ts
@@ -1,51 +1,43 @@
 import {
   addProgramMessage,
-  setShouldExit,
   setVerbose,
   forgetContextMessages,
-  setActiveView,
-  useChatState,
-  addLocalAiMessage,
-  addLocalUserMessage,
-} from './state.js';
+  showHelpView,
+  showInfoView,
+  triggerExit,
+} from './state/actions.js';
+import { useChatState } from './state/state.js';
 import { saveConversation } from './utils.js';
 
-export function processCommand(input: string) {
+export function processChatCommand(input: string) {
   if (!input.startsWith('/')) {
     return false;
   }
 
   const command = input.split(' ')[0];
   if (command === '/help') {
-    setActiveView('help');
+    showHelpView();
     return true;
   }
 
   if (command === '/info') {
-    setActiveView('info');
+    showInfoView();
     return true;
   }
 
-  addLocalUserMessage(input);
-
   const state = useChatState.getState();
-  setActiveView(null);
-
   if (command === '/exit') {
-    addLocalAiMessage('Bye! 👋');
-    setShouldExit(true);
+    triggerExit();
     return true;
   }
 
   if (command === '/verbose') {
     setVerbose(!state.verbose);
-    addProgramMessage(`Verbose mode: ${state.verbose ? 'off' : 'on'}`);
     return true;
   }
 
   if (command === '/forget') {
     forgetContextMessages();
-    addProgramMessage('AI will forget previous messages.');
     return true;
   }
