+++ src/commands/prompt/ui/list/UserChatMessageItem.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Text } from 'ink';
 import { colors } from '../../../../theme/colors.js';
-import { type UserChatMessage } from '../../state.js';
+import { type UserChatMessage } from '../../state/state.js';
 import { texts } from '../../texts.js';
 
 type UserChatMessageItemProps = {
